+++ application/uni-best-template/src/pages.json
@@ -63,18 +63,10 @@
       }
     },
     {
-      "path": "pages/appointment/form",
-      "type": "page",
-      "style": {
-        "navigationBarTitleText": "服务预约",
-        "navigationStyle": "custom"
-      }
-    },
-    {
       "path": "pages/appointment/index",
       "type": "page",
       "style": {
-        "navigationBarTitleText": "一键预约",
+        "navigationBarTitleText": "服务预约",
         "navigationStyle": "custom"
       }
     },
