+++ application/uni-best-template/src/pages/appointment/index.vue
@@ -1,297 +1,395 @@
 <route lang="json5">
 {
   style: {
-    navigationBarTitleText: '一键预约',
-    navigationStyle: 'custom',
+    navigationBarTitleText: '服务预约',
+    navigationStyle: 'custom'
   },
 }
 </route>
 
 <template>
-  <view class="appointment-container">
-    <view class="appointment-header">
-      <view class="header-title">
-        <text class="title-text">服务预约</text>
-      </view>
-      <text class="header-desc">三棵树防水修缮专业服务</text>
-    </view>
-    
-    <view class="appointment-form">
-      <view class="form-item">
-        <text class="form-label">服务类型</text>
-        <view class="service-type-selector">
-          <view 
-            class="type-item" 
-            :class="{ active: serviceType === 'repair' }"
-            @click="serviceType = 'repair'"
-          >
-            防水补漏
-          </view>
-          <view 
-            class="type-item" 
-            :class="{ active: serviceType === 'new' }"
-            @click="serviceType = 'new'"
-          >
-            新房防水施工
-          </view>
-        </view>
-      </view>
-      
-      <view class="form-item">
-        <text class="form-label">联系人</text>
-        <input class="input-field" placeholder="请输入您的姓名" v-model="formData.name" />
-      </view>
-      
-      <view class="form-item">
-        <text class="form-label">手机号码</text>
-        <input class="input-field" placeholder="请输入您的手机号" type="number" v-model="formData.phone" />
-      </view>
-      
-      <view class="form-item">
-        <text class="form-label">服务地址</text>
-        <input class="input-field" placeholder="请输入详细地址" v-model="formData.address" />
-      </view>
-      
-      <view class="form-item">
-        <text class="form-label">预约时间</text>
-        <picker mode="date" :start="startDate" :end="endDate" @change="onDateChange" class="date-picker">
-          <view class="picker-value">{{ formData.date || '请选择预约日期' }}</view>
-        </picker>
-      </view>
-      
-      <view class="form-item">
-        <text class="form-label">问题描述</text>
-        <textarea class="textarea-field" placeholder="请简要描述您的防水需求" v-model="formData.description" />
+  <view class="form-container">    
+    <view class="banner">
+      <view class="banner-text">
+        <view class="cn-text">中 国 奥 委 会</view>
+        <view class="supplier-text">官方涂料独家供应商</view>
       </view>
     </view>
     
-    <view class="submit-section">
-      <view class="submit-button" @click="submitAppointment">立即预约</view>
-      <view class="service-guarantee">
-        <text class="guarantee-text">服务承诺：1小时响应 | 专业团队 | 品质保障</text>
-      </view>
-    </view>
+    <wd-form :model="formData" align="right" ref="appointmentForm" validate-trigger="submit">
+      <wd-cell-group border>
+        <!-- 服务类型 -->
+        <wd-picker
+          label="选择服务类型"
+          label-width="140px"
+          prop="serviceType"
+          v-model="formData.serviceType"
+          :columns="serviceOptions"
+          placeholder="请选择服务类型"
+          :rules="[{ required: true, message: '请选择服务类型' }]"
+          custom-class="service-picker"
+        ></wd-picker>
+        
+        <!-- 场景类型 -->
+        <wd-select-picker
+          label="选择场景类型"
+          label-width="140px"
+          prop="sceneType"
+          v-model="formData.sceneType"
+          :columns="sceneOptions" 
+          mode="single-select"
+          placeholder="点击选择场景类型"
+          custom-class="scene-select"
+          :rules="[{ required: true, message: '请选择场景类型' }]"
+        ></wd-select-picker>
+      </wd-cell-group>
+      <wd-cell-group border>
+        <!-- 业主姓名 -->
+        <wd-input 
+          label="业主姓名"
+          label-width="140px"
+          prop="name"
+          v-model="formData.name" 
+          placeholder="请填写" 
+          custom-class="custom-input"
+          :rules="[{ required: true, message: '请输入业主姓名' }]"
+        >
+          <template #suffix>
+            <wd-icon name="user" size="36rpx" color="#999" v-if="!formData.name"></wd-icon>
+          </template>
+        </wd-input>
+        
+        <!-- 联系电话 -->
+        <wd-input 
+          label="联系电话"
+          label-width="140px"
+          prop="phone"
+          v-model="formData.phone" 
+          placeholder="请填写" 
+          type="number" 
+          custom-class="custom-input"
+          :rules="[
+            { required: true, message: '请输入联系电话' },
+            { pattern: /^1\d{10}$/, message: '请输入正确的手机号码' }
+          ]"
+        >
+          <template #suffix>
+            <wd-icon name="phone" size="36rpx" color="#999" v-if="!formData.phone"></wd-icon>
+          </template>
+        </wd-input>
+        
+        <!-- 小区位置 -->
+        <wd-input 
+          label="小区位置"
+          label-width="140px"
+          prop="location"
+          v-model="formData.location" 
+          placeholder="点击选择" 
+          readonly 
+          custom-class="custom-input"
+          @click="chooseLocation"
+          :rules="[{ required: true, message: '请选择小区位置' }]"
+        >
+          <template #suffix>
+            <wd-icon name="location" size="36rpx" color="#2c722c"></wd-icon>
+          </template>
+        </wd-input>
+        
+        <!-- 选择省市区 -->
+        <wd-input 
+          label="选择省市区"
+          label-width="140px"
+          prop="region"
+          v-model="formData.region" 
+          placeholder="请选择" 
+          readonly
+          custom-class="custom-input"
+          @click="chooseRegion"
+          :rules="[{ required: true, message: '请选择省市区' }]"
+        >
+          <template #suffix>
+            <wd-icon name="arrow-right" size="36rpx" color="#999"></wd-icon>
+          </template>
+        </wd-input>
+        
+        <!-- 详细地址 -->
+        <wd-input 
+          label="详细地址"
+          label-width="140px"
+          prop="address"
+          v-model="formData.address" 
+          placeholder="请填写" 
+          custom-class="custom-input"
+        >
+          <template #suffix>
+            <wd-icon name="edit-outline" size="36rpx" color="#999" v-if="!formData.address"></wd-icon>
+          </template>
+        </wd-input>
+      </wd-cell-group>
+    </wd-form>
     
-    <!-- 自定义tabbar -->
-    <fg-tabbar current="appointment"></fg-tabbar>
+    <!-- 添加底部占位，防止内容被提交按钮遮挡 -->
+    <view style="height: 160rpx;"></view>
+    
+    <view class="submit-wrapper">
+      <wd-button size="large" type="primary" block custom-class="submit-btn" @click="submitForm">确认预约</wd-button>
+      <view class="safe-area-inset-bottom"></view>
+    </view>
   </view>
 </template>
 
 <script lang="ts" setup>
-import { ref, reactive, computed } from 'vue'
-import FgTabbar from '@/components/fg-tabbar/fg-tabbar.vue'
+import { ref, reactive, onMounted } from 'vue'
 
-// 服务类型
-const serviceType = ref('repair')
+// 安全距离
+const safeAreaBottom = ref(0)
+
+// 表单引用
+const appointmentForm = ref()
+
+onMounted(() => {
+  // 获取安全区域信息
+  uni.getSystemInfo({
+    success: (res) => {
+      console.log('系统信息', res)
+      if (res.safeAreaInsets) {
+        safeAreaBottom.value = res.safeAreaInsets.bottom || 0
+      }
+    }
+  })
+})
 
 // 表单数据
 const formData = reactive({
+  serviceType: '',
   name: '',
   phone: '',
+  region: '',
   address: '',
-  date: '',
-  description: ''
+  sceneType: '',
+  location: ''
 })
 
-// 预约日期范围
-const startDate = computed(() => {
-  const today = new Date()
-  return today.toISOString().split('T')[0]
-})
+// 服务类型选项
+const serviceOptions = ref([
+  { value: 'repair', label: '防水补漏' },
+  { value: 'new', label: '新房防水施工' },
+  { value: 'unsure', label: '我不清楚' }
+])
 
-const endDate = computed(() => {
-  const today = new Date()
-  today.setMonth(today.getMonth() + 3)
-  return today.toISOString().split('T')[0]
-})
+// 场景选项数据
+const sceneOptions = ref([
+  { value: '客厅', label: '客厅' },
+  { value: '厨房', label: '厨房' },
+  { value: '卫生间', label: '卫生间' },
+  { value: '阳台', label: '阳台' },
+  { value: '卧室', label: '卧室' },
+  { value: '其他', label: '其他' }
+])
 
-// 日期选择
-const onDateChange = (e: any) => {
-  formData.date = e.detail.value
+// 选择位置
+const chooseLocation = () => {
+  // #ifdef MP-WEIXIN
+  uni.chooseLocation({
+    success: (res) => {
+      console.log('选择位置成功', res)
+      // 自动填充省市区和详细地址
+      formData.location = res.name || res.address
+      formData.region = res.address.split(',')[0] || ''
+      formData.address = res.address || ''
+    },
+    fail: (err) => {
+      console.log('选择位置失败', err)
+      uni.showToast({
+        title: '请授权位置权限',
+        icon: 'none'
+      })
+    }
+  })
+  // #endif
+  
+  // #ifndef MP-WEIXIN
+  uni.showToast({
+    title: '仅微信小程序支持此功能',
+    icon: 'none'
+  })
+  // #endif
 }
 
-// 提交预约
-const submitAppointment = () => {
-  // 表单验证
-  if (!formData.name) {
-    return uni.showToast({ title: '请输入联系人', icon: 'none' })
-  }
-  
-  if (!formData.phone) {
-    return uni.showToast({ title: '请输入手机号码', icon: 'none' })
-  }
-  
-  if (!/^1\d{10}$/.test(formData.phone)) {
-    return uni.showToast({ title: '请输入正确的手机号码', icon: 'none' })
-  }
-  
-  if (!formData.address) {
-    return uni.showToast({ title: '请输入服务地址', icon: 'none' })
-  }
-  
-  if (!formData.date) {
-    return uni.showToast({ title: '请选择预约时间', icon: 'none' })
-  }
-  
-  // 提交表单
-  uni.showLoading({ title: '提交中...' })
-  
-  setTimeout(() => {
-    uni.hideLoading()
-    uni.showToast({
-      title: '预约成功，我们将尽快联系您',
-      icon: 'none',
-      duration: 2000
+// 选择省市区
+const chooseRegion = () => {
+  // 在微信环境中使用地图选择器
+  if (formData.location) {
+    return uni.showToast({ 
+      title: '省市区已通过位置选择器自动填写', 
+      icon: 'none' 
     })
-    
-    // 重置表单
-    formData.name = ''
-    formData.phone = ''
-    formData.address = ''
-    formData.date = ''
-    formData.description = ''
-    
-    // 跳转到首页
-    setTimeout(() => {
-      uni.switchTab({ url: '/pages/index/index' })
-    }, 2000)
-  }, 1500)
+  }
+  
+  // 提示用户先选择位置
+  uni.showToast({ 
+    title: '请先选择小区位置，省市区将自动填写', 
+    icon: 'none' 
+  })
+}
+
+// 提交表单
+const submitForm = () => {
+  // 使用wd-form进行表单验证
+  appointmentForm.value.validate().then(({ valid, errors }) => {
+    if (valid) {
+      // 表单验证通过，提交数据
+      uni.showLoading({ title: '提交中...' })
+      
+      setTimeout(() => {
+        uni.hideLoading()
+        uni.showToast({
+          title: '预约成功，我们将尽快联系您',
+          icon: 'none',
+          duration: 2000
+        })
+        
+        // 跳转到首页
+        setTimeout(() => {
+          uni.switchTab({ url: '/pages/index/index' })
+        }, 2000)
+      }, 1500)
+    } else {
+      console.log('表单验证失败', errors)
+    }
+  }).catch(error => {
+    console.log('验证出错', error)
+  })
 }
 </script>
 
-<style lang="scss">
-.appointment-container {
+<style lang="scss" scoped>
+.form-container {
   display: flex;
   flex-direction: column;
-  min-height: 100vh;
-  background-color: #f5f9f5;
-  padding-bottom: 100rpx;
+  // min-height: 100vh;
+  background-color: #f7f8fa;
+  position: relative;
 }
 
-.appointment-header {
-  padding: 60rpx 40rpx 40rpx;
-  background-color: #2c722c;
+.banner {
+  position: relative;
+  width: 100%;
+  height: 140rpx;
+  background: linear-gradient(to right, #3d9c40, #67c541);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 20rpx;
   
-  .header-title {
-    margin-bottom: 20rpx;
-    
-    .title-text {
-      font-size: 48rpx;
-      font-weight: bold;
-      color: #fff;
-    }
-  }
-  
-  .header-desc {
-    font-size: 28rpx;
-    color: rgba(255, 255, 255, 0.8);
-  }
-}
-
-.appointment-form {
-  padding: 30rpx;
-  background-color: #fff;
-  margin: 30rpx;
-  border-radius: 16rpx;
-  box-shadow: 0 4rpx 20rpx rgba(0, 0, 0, 0.05);
-  
-  .form-item {
-    margin-bottom: 40rpx;
-    
-    .form-label {
-      display: block;
-      font-size: 28rpx;
-      color: #333;
-      margin-bottom: 20rpx;
-      font-weight: bold;
-    }
-    
-    .input-field {
-      width: 100%;
-      height: 80rpx;
-      border-radius: 8rpx;
-      background-color: #f8f8f8;
-      padding: 0 20rpx;
-      box-sizing: border-box;
-      font-size: 28rpx;
-    }
-    
-    .textarea-field {
-      width: 100%;
-      height: 200rpx;
-      border-radius: 8rpx;
-      background-color: #f8f8f8;
-      padding: 20rpx;
-      box-sizing: border-box;
-      font-size: 28rpx;
-    }
-    
-    .service-type-selector {
-      display: flex;
-      justify-content: space-between;
-      
-      .type-item {
-        width: 48%;
-        height: 80rpx;
-        border-radius: 8rpx;
-        background-color: #f8f8f8;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        font-size: 28rpx;
-        color: #666;
-        
-        &.active {
-          background-color: #e8f4e8;
-          color: #2c722c;
-          font-weight: bold;
-          border: 2rpx solid #2c722c;
-        }
-      }
-    }
-    
-    .date-picker {
-      width: 100%;
-      height: 80rpx;
-      border-radius: 8rpx;
-      background-color: #f8f8f8;
-      padding: 0 20rpx;
-      box-sizing: border-box;
-      
-      .picker-value {
-        height: 80rpx;
-        line-height: 80rpx;
-        font-size: 28rpx;
-        color: #333;
-      }
-    }
-  }
-}
-
-.submit-section {
-  padding: 0 30rpx;
-  margin-top: 30rpx;
-  
-  .submit-button {
-    height: 90rpx;
-    border-radius: 45rpx;
-    background-color: #2c722c;
-    color: #fff;
-    font-size: 32rpx;
-    font-weight: bold;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  .service-guarantee {
-    margin-top: 30rpx;
+  .banner-text {
     text-align: center;
+    color: #fff;
     
-    .guarantee-text {
-      font-size: 24rpx;
-      color: #999;
+    .cn-text {
+      font-size: 36rpx;
+      letter-spacing: 12rpx;
+      margin-bottom: 10rpx;
+      font-weight: bold;
     }
+    
+    .supplier-text {
+      font-size: 30rpx;
+    }
+  }
+}
+
+.service-options {
+  display: flex;
+  width: 100%;
+  // justify-content: flex-end;
+}
+
+.radio-group {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  
+  :deep(.wd-radio) {
+    margin-bottom: 15rpx;
+    margin-right: 0;
+  }
+  
+  :deep(.wd-radio__shape) {
+    margin-right: 10rpx;
+  }
+  
+  :deep(.wd-radio__button) {
+    width: 100%;
+    height: 80rpx;
+    text-align: right;
+    padding-right: 20rpx;
+    border-radius: 10rpx;
+    
+    &.is-checked {
+      background-color: #e8f5e9;
+      color: #2c722c;
+      border-color: #2c722c;
+    }
+  }
+}
+
+.custom-cell {
+  :deep(.wd-cell__title) {
+    font-weight: 500;
+  }
+}
+
+.scene-select {
+  :deep(.wd-picker__value) {
+    text-align: right;
+  }
+}
+
+:deep(.wd-cell-group) {
+  margin: 0 30rpx;
+  border-radius: 8rpx;
+  overflow: hidden;
+}
+
+:deep(.wd-cell__title) {
+  &::before {
+    content: '*';
+    color: #ff4d4f;
+    position: absolute;
+    left: -15rpx;
+  }
+}
+
+:deep(.wd-input__inner) {
+  // text-align: right;
+}
+
+.submit-wrapper {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  padding: 30rpx;
+  z-index: 9;
+}
+
+.submit-btn {
+  height: 90rpx;
+  font-size: 32rpx;
+  font-weight: bold;
+  border-radius: 45rpx;
+  background-color: #3d9c40 !important;
+  border-color: #3d9c40 !important;
+}
+
+.safe-area-inset-bottom {
+  height: v-bind('safeAreaBottom + "px"');
+}
+
+.service-picker {
+  :deep(.wd-picker__value) {
+    text-align: right;
   }
 }
 </style>
